Moved item instances to subpackage for items. Renamed item subpackage 'blocks' to 'tiles' to be more inclusive of walls.
