+++ src/main/java/org/egordorichev/lasttry/entity/Entity.java
@@ -1,7 +1,7 @@
 package org.egordorichev.lasttry.entity;
 
 import org.egordorichev.lasttry.LastTry;
-import org.egordorichev.lasttry.item.Block;
+import org.egordorichev.lasttry.item.tiles.Block;
 import org.egordorichev.lasttry.util.Direction;
 import org.egordorichev.lasttry.util.Rectangle;
 import org.newdawn.slick.Image;
