+++ src/main/java/org/egordorichev/lasttry/item/Item.java
@@ -1,7 +1,9 @@
 package org.egordorichev.lasttry.item;
 
 import org.egordorichev.lasttry.LastTry;
-import org.egordorichev.lasttry.item.blocks.BlockGround;
+import org.egordorichev.lasttry.item.items.Coin;
+import org.egordorichev.lasttry.item.tiles.BlockGround;
+import org.egordorichev.lasttry.item.tiles.Wall;
 import org.egordorichev.lasttry.util.Assets;
 import org.newdawn.slick.Image;
