+++ src/main/java/org/egordorichev/lasttry/item/items/Accessory.java
@@ -1,5 +1,6 @@
-package org.egordorichev.lasttry.item;
+package org.egordorichev.lasttry.item.items;
 
+import org.egordorichev.lasttry.item.Item;
 import org.newdawn.slick.Image;
 
 public class Accessory extends Item {
