+++ src/main/java/org/egordorichev/lasttry/item/tiles/Block.java
@@ -1,6 +1,7 @@
-package org.egordorichev.lasttry.item;
+package org.egordorichev.lasttry.item.tiles;
 
 import org.egordorichev.lasttry.LastTry;
+import org.egordorichev.lasttry.item.Item;
 import org.egordorichev.lasttry.world.tile.TileData;
 import org.newdawn.slick.Image;
