+++ src/main/java/org/egordorichev/lasttry/item/tiles/MultiTileBlock.java
@@ -1,6 +1,5 @@
-package org.egordorichev.lasttry.item.blocks;
+package org.egordorichev.lasttry.item.tiles;
 
-import org.egordorichev.lasttry.item.Block;
 import org.egordorichev.lasttry.world.tile.TileData;
 import org.newdawn.slick.Image;
